Added Product and Mean to Univariate and all implementations, this patch contains contributions from Mark Diggory.
* This patch introduces Product and GeometricMean into the Univariate
implementation.

* Discarding the contribution of a discarded element in a rolling
UnivariateImpl requires that the product be calculated explicitly each
time a value is discarded.  This is necessary because not all values will
have non-zero values.

* Errors in rolling logic for ListUimpl, and UnivariateImpl were corrected,
and more test cases were added to the JUnit tests for the Univariate
implementations.  More rigorous test cases are needed for the entire
suite of Univariate implementations


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/math/trunk@140851 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/commons/math/StoreUnivariateImpl.java
@@ -91,7 +91,7 @@ public class StoreUnivariateImpl extends AbstractStoreUnivariate {
     /* (non-Javadoc)
      * @see org.apache.commons.math.Univariate#getN()
      */
-    public double getN() {
+    public int getN() {
         return eDA.getNumElements();
     }
